+++ signing/secp256k1/api/src/main/java/tech/pegasys/signers/secp256k1/api/SingleTransactionSignerProvider.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.signers.secp256k1.signerapi;
+package tech.pegasys.signers.secp256k1.api;
 
 import java.util.Collections;
 import java.util.Optional;
