+++ signing/secp256k1/azure/src/main/java/tech/pegasys/signers/secp256k1/azure/AzureKeyVaultTransactionSignerFactory.java
@@ -14,8 +14,8 @@ package tech.pegasys.signers.secp256k1.azure;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
+import tech.pegasys.signers.secp256k1.api.TransactionSigner;
 import tech.pegasys.signers.secp256k1.multikey.TransactionSignerInitializationException;
-import tech.pegasys.signers.secp256k1.signerapi.TransactionSigner;
 
 import java.math.BigInteger;
 import java.net.UnknownHostException;
