+++ signing/secp256k1/azure/src/test/java/tech/pegasys/signers/secp256k1/azure/AzureKeyVaultTransactionSignerProviderTest.java
@@ -17,7 +17,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import tech.pegasys.signers.secp256k1.signerapi.TransactionSigner;
+import tech.pegasys.signers.secp256k1.api.TransactionSigner;
 
 import java.math.BigInteger;
 import java.util.Arrays;
