+++ signing/secp256k1/file-based/src/test/java/tech/pegasys/signers/secp256k1/filebased/FileBasedTransactionSignerTest.java
@@ -16,8 +16,8 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import tech.pegasys.signers.secp256k1.api.TransactionSigner;
 import tech.pegasys.signers.secp256k1.multikey.TransactionSignerInitializationException;
-import tech.pegasys.signers.secp256k1.signerapi.TransactionSigner;
 
 import java.io.File;
 import java.io.IOException;
