+++ signing/secp256k1/multikey/src/main/java/tech/pegasys/signers/secp256k1/multikey/MultiSignerFactory.java
@@ -12,10 +12,10 @@
  */
 package tech.pegasys.signers.secp256k1.multikey;
 
+import tech.pegasys.signers.secp256k1.api.TransactionSigner;
 import tech.pegasys.signers.secp256k1.multikey.metadata.AzureSigningMetadataFile;
 import tech.pegasys.signers.secp256k1.multikey.metadata.FileBasedSigningMetadataFile;
 import tech.pegasys.signers.secp256k1.multikey.metadata.HashicorpSigningMetadataFile;
-import tech.pegasys.signers.secp256k1.signerapi.TransactionSigner;
 
 public interface MultiSignerFactory {
