+++ signing/secp256k1/multikey/src/main/java/tech/pegasys/signers/secp256k1/multikey/metadata/SigningMetadataFile.java
@@ -12,8 +12,8 @@
  */
 package tech.pegasys.signers.secp256k1.multikey.metadata;
 
+import tech.pegasys.signers.secp256k1.api.TransactionSigner;
 import tech.pegasys.signers.secp256k1.multikey.MultiSignerFactory;
-import tech.pegasys.signers.secp256k1.signerapi.TransactionSigner;
 
 public abstract class SigningMetadataFile {
